moved parameter definition to separate file to be called by detect script
